+++ Lunar/Settings Views/SettingsDevOptionsView.swift
@@ -6,7 +6,6 @@
 //
 
 import Defaults
-import LocalConsole
 import Pulse
 import PulseUI
 import SFSafeSymbols
@@ -22,33 +21,12 @@ struct SettingsDevOptionsView: View {
   @Default(.accentColor) var accentColor
 
   @State var settingsViewOpacity: Double = 1
-  @State private var localConsoleButtonText: String = "Enable Local Console"
 
   let notificationHaptics = UINotificationFeedbackGenerator()
   let haptics = UIImpactFeedbackGenerator(style: .soft)
 
   var body: some View {
     List {
-      // MARK: - LOCAL CONSOLE
-
-      Section {
-        Button {
-          haptics.impactOccurred(intensity: 0.7)
-          LCManager.shared.isVisible.toggle()
-          localConsoleButtonText =
-            LCManager.shared.isVisible ? "Close Local Console" : "Enable Local Console"
-
-        } label: {
-          Label {
-            Text(localConsoleButtonText)
-              .foregroundStyle(.foreground)
-          } icon: {
-            Image(systemSymbol: .macwindow)
-              .symbolRenderingMode(.multicolor)
-          }
-        }
-      }
-
       // MARK: - DEBUG AND PULSE
 
       Section {
@@ -151,30 +129,6 @@ struct SettingsDevOptionsView: View {
               .symbolRenderingMode(.hierarchical)
           }
         }
-
-        Toggle(isOn: $realmExperimentalViewEnabled) {
-          Label {
-            Text("Realm Experimental View")
-          } icon: {
-            Image(systemSymbol: .circleAndLineHorizontalFill)
-              .foregroundStyle(.pink)
-              .symbolRenderingMode(.hierarchical)
-          }
-        }
-        .tint(accentColor)
-
-        //        NavigationLink {
-        //          OfflineDownloaderView()
-        //        } label: {
-        //          Label {
-        //            Text("Offline Downloader")
-        //          } icon: {
-        //            Image(systemSymbol: .squareAndArrowDownFill)
-        //              .foregroundStyle(.cyan)
-        //              .brightness(-0.2)
-        //              .symbolRenderingMode(.hierarchical)
-        //          }
-        //        }
 
         NavigationLink {
           ColorTesterView()